First step in APIs integration
SEARCH PANEL
People
The typeahead for people is not ready yet. Mainly for this reason
nothing has been done yet on the individual network.
On the shared network: connected the http request for shared network
with manually inserted IDs (see line 35 in searchpanel.js).
The endpoint is not working with all the ids (ex not working with
shakespeare + milton, ids = [10010937,10008309])

Groups
Only thing done: implemented the singular group typeahead. It gets
groups name, you select one and it call for further data. Single group
network api currently not working.